Build tag to exclude ArchiverProvider
